Refactor update_zone_value so the logic that is specific to DBA, and cannot be reused by Triumph resides in its own file.
Triumph! will have the same change made moving its specfic functions
to its own file.

The idea is to make it easier to move code between DBA and Triumph,
and vice versa, by having the specific code in its own files.

Then when pulling from the other repo, it is easy to see what is generic
code and accept it.