Refactor update_zone_value to use extension points.
DBA and Triumph each have their own definition of the extension points
for calculating the dead, and which skeleton should be shown based
on how many dead there is.

The idea is to make it easier to move code between DBA and Triumph,
and vice versa, by having the specific code in its own files.

cherry-picking commit 7e2681d.