Copied adding of buildings from buildings_squashed branch in DBA
When the table is locked buildings are added to the Villages to make the
table look pretty.

When a base is approaching the building, the building is destroyed.
When all bases leave the area the build is spawned again.

In Triumph! the Village area is important, not the buildnig itself.

The building location is saved and restored with the game.  If a
base is located where a building was, and the game is saved,
loaded, and the base moves out the building will be retored.

The definitions of the buildiings comes from data_terrain.ttslua